Inline GdbOsReaderOptions.
